+++ CHANGELOG.md
@@ -118,6 +118,9 @@ format.
 - Rewrite article/profile social metadata page URLs in static exports,
   including Open Graph related links and Twitter player pages.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
+- Copy dependencies discovered inside linked copied text assets, including
+  icons referenced by web manifests.
+  ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
 
 ### Documentation